Allow dashes in acronyms
Sometimes acronyms might be combined with numbers such as `IEC-8859-1`
and are considered a problem. This patch adds dashes to the charset that
acronyms can be made of and therefore should get rid of the matching.

One could argue that with dashes it might no longer technically be a
sole acronym but at least in a technical context I think it's quite
common to mix those and it would be quite nice to not being in the need
of adding an explicit excemption for each of them.